Fix horrific echo logic in app code
Since the app uses UDP to communicate with the console (for the
ability to broadcast), the net logic checks for echoes when waiting
for a response. The backwards logic I has converted the addresses to
strings and did a "contains" check, so if the console was .4 and the
app was .42, it was considered an echo... This patch compares the
addresses properly.